+++ lib/src/core/entity_system.dart
@@ -125,4 +125,6 @@ abstract class EntitySystem implements EntityObserver {
     }
   }
 
+  void destroy() {}
+
 }